+++ src/Board.h
@@ -21,16 +21,16 @@
 			static void reset();
 			static void doMove(uint32_t move);
 			static void undoMove();
-			static uint8_t turn;
-			static void put(bool color, uint8_t piece, uint8_t location);
-			static void remove(bool color, uint8_t piece, uint8_t location);
-			static uint8_t getPiece(uint8_t location);
-			static uint8_t getColor(uint8_t location);
+			static uint32_t turn;
+			static void put(bool color, uint32_t piece, uint32_t location);
+			static void remove(bool color, uint32_t piece, uint32_t location);
+			static uint32_t getPiece(uint32_t location);
+			static uint32_t getColor(uint32_t location);
 			static uint64_t occupiedSquares;
 			static uint64_t colors[2];
 			static uint64_t pieces[8];
-			static uint8_t piecesIndex[64];
-			static uint8_t colorsIndex[64];
+			static uint32_t piecesIndex[64];
+			static uint32_t colorsIndex[64];
 
 	};
 #endif
